+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 HealthModule modules from being installed by BeadledomModule. If the removed functionality is 
 desired, install the removed modules in the consuming guice module.
 * Bump minimum Java version to 1.8 for all modules.
+* Change the HttpClient `ServiceUnavailableRetryStrategy` to only retry on 503 response codes.
 
 ### Enhancements
 * Support building with JDK 9/10+